requirements.txt
pyserial==3.4

matplotlib

pypiwin32

tweepy 
